(PC-26282)[PRO] refactor: remove intermediary useState and useEffect
